Extract release name and namespace from input
The ResourceList kind has a field for functionConfig, which can be
used to supply configuration to the container image. This commit adds
code to the run-helm-sh script to extract that config and get
Helm-specific values from it.

This will come into play if you do, for example,

   kpt fn run instance --image helm-flux -- releaseName=flux

(the trailing arguments are turned into a ConfigMap and put in the
functionConfig slot by `kpt fn run`)