Make template generation replace files
When you change e.g., the release name, and rerun the chart templates,
the files will get different names. If the existing files are included
in the output, there will be near duplication of the resources, which
is not very useful.

This commit changes the script so it removes the set of existing files
before running the chart. To be able to do this it has to keep the
generated files separate.